+++ example/test/widget_test.dart
@@ -19,8 +19,6 @@ void main() {
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
     expect(find.text('1'), findsNothing);
-    expect(find.text('Tapped 0 times'), findsOneWidget);
-    expect(find.text('Tapped 1 times'), findsNothing);
 
     // Tap the '+' icon and trigger a frame.
     await tester.tap(find.byIcon(Icons.add));
@@ -29,7 +27,5 @@ void main() {
     // Verify that our counter has incremented.
     expect(find.text('0'), findsNothing);
     expect(find.text('1'), findsOneWidget);
-    expect(find.text('Tapped 0 times'), findsNothing);
-    expect(find.text('Tapped 1 times'), findsOneWidget);
   });
 }