+++ lib/src/change_notifier_provider.dart
@@ -108,12 +108,14 @@ class ChangeNotifierProvider<T extends ChangeNotifier> extends ListenableProvide
     Key key,
     @required Create<T> create,
     bool lazy,
+    ValueWidgetBuilder<T> builder,
     Widget child,
   }) : super(
           key: key,
           create: create,
           dispose: _dispose,
           lazy: lazy,
+          builder: builder,
           child: child,
         );
 
@@ -121,9 +123,11 @@ class ChangeNotifierProvider<T extends ChangeNotifier> extends ListenableProvide
   ChangeNotifierProvider.value({
     Key key,
     @required T value,
+    ValueWidgetBuilder<T> builder,
     Widget child,
   }) : super.value(
           key: key,
+          builder: builder,
           value: value,
           child: child,
         );
@@ -210,6 +214,7 @@ class ChangeNotifierProxyProvider<T, R extends ChangeNotifier> extends Listenabl
     @required Create<R> create,
     @required ProxyProviderBuilder<T, R> update,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   }) : super(
           key: key,
@@ -217,6 +222,7 @@ class ChangeNotifierProxyProvider<T, R extends ChangeNotifier> extends Listenabl
           update: update,
           dispose: ChangeNotifierProvider._dispose,
           lazy: lazy,
+          builder: builder,
           child: child,
         );
 }
@@ -229,6 +235,7 @@ class ChangeNotifierProxyProvider0<R extends ChangeNotifier> extends ListenableP
     @required Create<R> create,
     @required R Function(BuildContext, R value) update,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   }) : super(
           key: key,
@@ -236,6 +243,7 @@ class ChangeNotifierProxyProvider0<R extends ChangeNotifier> extends ListenableP
           update: update,
           dispose: ChangeNotifierProvider._dispose,
           lazy: lazy,
+          builder: builder,
           child: child,
         );
 }
@@ -248,6 +256,7 @@ class ChangeNotifierProxyProvider2<T, T2, R extends ChangeNotifier> extends List
     @required Create<R> create,
     @required ProxyProviderBuilder2<T, T2, R> update,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   }) : super(
           key: key,
@@ -255,6 +264,7 @@ class ChangeNotifierProxyProvider2<T, T2, R extends ChangeNotifier> extends List
           update: update,
           dispose: ChangeNotifierProvider._dispose,
           lazy: lazy,
+          builder: builder,
           child: child,
         );
 }
@@ -267,6 +277,7 @@ class ChangeNotifierProxyProvider3<T, T2, T3, R extends ChangeNotifier> extends 
     @required Create<R> create,
     @required ProxyProviderBuilder3<T, T2, T3, R> update,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   }) : super(
           key: key,
@@ -274,6 +285,7 @@ class ChangeNotifierProxyProvider3<T, T2, T3, R extends ChangeNotifier> extends 
           update: update,
           dispose: ChangeNotifierProvider._dispose,
           lazy: lazy,
+          builder: builder,
           child: child,
         );
 }
@@ -287,6 +299,7 @@ class ChangeNotifierProxyProvider4<T, T2, T3, T4, R extends ChangeNotifier>
     @required Create<R> create,
     @required ProxyProviderBuilder4<T, T2, T3, T4, R> update,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   }) : super(
           key: key,
@@ -294,6 +307,7 @@ class ChangeNotifierProxyProvider4<T, T2, T3, T4, R extends ChangeNotifier>
           update: update,
           dispose: ChangeNotifierProvider._dispose,
           lazy: lazy,
+          builder: builder,
           child: child,
         );
 }
@@ -307,6 +321,7 @@ class ChangeNotifierProxyProvider5<T, T2, T3, T4, T5, R extends ChangeNotifier>
     @required Create<R> create,
     @required ProxyProviderBuilder5<T, T2, T3, T4, T5, R> update,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   }) : super(
           key: key,
@@ -314,6 +329,7 @@ class ChangeNotifierProxyProvider5<T, T2, T3, T4, T5, R extends ChangeNotifier>
           update: update,
           dispose: ChangeNotifierProvider._dispose,
           lazy: lazy,
+          builder: builder,
           child: child,
         );
 }
@@ -327,6 +343,7 @@ class ChangeNotifierProxyProvider6<T, T2, T3, T4, T5, T6, R extends ChangeNotifi
     @required Create<R> create,
     @required ProxyProviderBuilder6<T, T2, T3, T4, T5, T6, R> update,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   }) : super(
           key: key,
@@ -334,6 +351,7 @@ class ChangeNotifierProxyProvider6<T, T2, T3, T4, T5, T6, R extends ChangeNotifi
           update: update,
           dispose: ChangeNotifierProvider._dispose,
           lazy: lazy,
+          builder: builder,
           child: child,
         );
 }