+++ lib/src/listenable_provider.dart
@@ -26,6 +26,7 @@ class ListenableProvider<T extends Listenable> extends InheritedProvider<T> {
     @required Create<T> create,
     Dispose<T> dispose,
     bool lazy,
+    ValueWidgetBuilder<T> builder,
     Widget child,
   })  : assert(create != null),
         super(
@@ -63,6 +64,7 @@ ChangeNotifierProvider(
                   }
                 },
           lazy: lazy,
+          builder: builder,
           child: child,
         );
 
@@ -71,9 +73,11 @@ ChangeNotifierProvider(
     Key key,
     @required T value,
     UpdateShouldNotify<T> updateShouldNotify,
+    ValueWidgetBuilder<T> builder,
     Widget child,
   }) : super.value(
           key: key,
+          builder: builder,
           value: value,
           updateShouldNotify: updateShouldNotify,
           startListening: _startListening,
@@ -99,6 +103,7 @@ class ListenableProxyProvider0<R extends Listenable> extends InheritedProvider<R
     Dispose<R> dispose,
     UpdateShouldNotify<R> updateShouldNotify,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   })  : assert(create != null || update != null),
         super(
@@ -106,6 +111,7 @@ class ListenableProxyProvider0<R extends Listenable> extends InheritedProvider<R
           create: create,
           update: update,
           lazy: lazy,
+          builder: builder,
           dispose: dispose,
           updateShouldNotify: updateShouldNotify,
           startListening: ListenableProvider._startListening,
@@ -143,12 +149,14 @@ class ListenableProxyProvider<T, R extends Listenable> extends ListenableProxyPr
     @required ProxyProviderBuilder<T, R> update,
     Dispose<R> dispose,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   })  : assert(create != null || update != null),
         super(
           key: key,
           create: create,
           lazy: lazy,
+          builder: builder,
           update: (context, previous) => update(
             context,
             Provider.of(context),
@@ -168,12 +176,14 @@ class ListenableProxyProvider2<T, T2, R extends Listenable> extends ListenablePr
     @required ProxyProviderBuilder2<T, T2, R> update,
     Dispose<R> dispose,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   })  : assert(create != null || update != null),
         super(
           key: key,
           create: create,
           lazy: lazy,
+          builder: builder,
           update: (context, previous) => update(
             context,
             Provider.of(context),
@@ -194,12 +204,14 @@ class ListenableProxyProvider3<T, T2, T3, R extends Listenable> extends Listenab
     @required ProxyProviderBuilder3<T, T2, T3, R> update,
     Dispose<R> dispose,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   })  : assert(create != null || update != null),
         super(
           key: key,
           create: create,
           lazy: lazy,
+          builder: builder,
           update: (context, previous) => update(
             context,
             Provider.of(context),
@@ -221,12 +233,14 @@ class ListenableProxyProvider4<T, T2, T3, T4, R extends Listenable> extends List
     @required ProxyProviderBuilder4<T, T2, T3, T4, R> update,
     Dispose<R> dispose,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   })  : assert(create != null || update != null),
         super(
           key: key,
           create: create,
           lazy: lazy,
+          builder: builder,
           update: (context, previous) => update(
             context,
             Provider.of(context),
@@ -249,12 +263,14 @@ class ListenableProxyProvider5<T, T2, T3, T4, T5, R extends Listenable> extends 
     @required ProxyProviderBuilder5<T, T2, T3, T4, T5, R> update,
     Dispose<R> dispose,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   })  : assert(create != null || update != null),
         super(
           key: key,
           create: create,
           lazy: lazy,
+          builder: builder,
           update: (context, previous) => update(
             context,
             Provider.of(context),
@@ -278,12 +294,14 @@ class ListenableProxyProvider6<T, T2, T3, T4, T5, T6, R extends Listenable> exte
     @required ProxyProviderBuilder6<T, T2, T3, T4, T5, T6, R> update,
     Dispose<R> dispose,
     bool lazy,
+    ValueWidgetBuilder<R> builder,
     Widget child,
   })  : assert(create != null || update != null),
         super(
           key: key,
           create: create,
           lazy: lazy,
+          builder: builder,
           update: (context, previous) => update(
             context,
             Provider.of(context),