+++ test/context_test.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/widgets.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:provider/provider.dart';
+
+void main() {
+  group('BuildContext', () {
+    testWidgets('context.read does not listen to value changes', (tester) async {
+      final child = Builder(builder: (context) {
+        final value = context.read<int>();
+        return Text('$value', textDirection: TextDirection.ltr);
+      });
+
+      await tester.pumpWidget(
+        Provider.value(
+          value: 42,
+          child: child,
+        ),
+      );
+
+      expect(find.text('42'), findsOneWidget);
+
+      await tester.pumpWidget(
+        Provider.value(
+          value: 24,
+          child: child,
+        ),
+      );
+
+      expect(find.text('42'), findsOneWidget);
+    });
+    testWidgets('context.watch listens to value changes', (tester) async {
+      final child = Builder(builder: (context) {
+        final value = context.watch<int>();
+        return Text('$value', textDirection: TextDirection.ltr);
+      });
+
+      await tester.pumpWidget(
+        Provider.value(
+          value: 42,
+          child: child,
+        ),
+      );
+
+      expect(find.text('42'), findsOneWidget);
+
+      await tester.pumpWidget(
+        Provider.value(
+          value: 24,
+          child: child,
+        ),
+      );
+
+      expect(find.text('24'), findsOneWidget);
+    });
+  });
+}