+++ test/inherited_provider_test.dart
@@ -18,7 +18,9 @@ BuildContext get context => find.byType(Context).evaluate().single;
 T of<T>([BuildContext c]) => Provider.of<T>(c ?? context, listen: false);
 
 void main() {
-  testWidgets('Provider.of(listen: false) outside of build works when it loads a provider', (tester) async {
+  testWidgets(
+      'Provider.of(listen: false) outside of build works when it loads a provider',
+      (tester) async {
     final notifier = ValueNotifier(42);
     await tester.pumpWidget(
       MultiProvider(
@@ -37,7 +39,8 @@ void main() {
 
     expect(Provider.of<String>(context, listen: false), '21');
   });
-  testWidgets('new value is available in didChangeDependencies', (tester) async {
+  testWidgets('new value is available in didChangeDependencies',
+      (tester) async {
     final didChangeDependencies = ValueBuilderMock<int>();
     final build = ValueBuilderMock<int>();
 
@@ -70,7 +73,65 @@ void main() {
     verifyNoMoreInteractions(didChangeDependencies);
     verifyNoMoreInteractions(build);
   });
-  testWidgets('provider.of throws if listen:true outside of the widget tree', (tester) async {
+  testWidgets(
+      'builder receives the current value and updates independently from `update`',
+      (tester) async {
+    final child = Container();
+
+    var notifier = ValueNotifier(0);
+    final update = ValueBuilderMock(0);
+    final builder = ValueWidgetBuilderMock<int>((c, value, child) {
+      final notifier = Provider.of<ValueNotifier<int>>(c);
+      return Text(
+        'notifier: ${notifier.value} value: $value',
+        textDirection: TextDirection.ltr,
+      );
+    });
+
+    await tester.pumpWidget(
+      ChangeNotifierProvider(
+        create: (_) => notifier,
+        child: InheritedProvider<int>(
+          update: update,
+          builder: builder,
+          child: child,
+        ),
+      ),
+    );
+
+    expect(find.text('notifier: 0 value: 0'), findsOneWidget);
+    verify(builder(argThat(isNotNull), 0, child)).called(1);
+    verifyNoMoreInteractions(builder);
+    verify(update(any, any)).called(1);
+    verifyNoMoreInteractions(update);
+
+    notifier.value++;
+    await tester.pump();
+
+    expect(find.text('notifier: 1 value: 0'), findsOneWidget);
+    verify(builder(argThat(isNotNull), 0, child)).called(1);
+    verifyNoMoreInteractions(builder);
+    verifyNoMoreInteractions(update);
+
+    when(update(any, any)).thenReturn(1);
+    await tester.pumpWidget(
+      ChangeNotifierProvider(
+        create: (_) => notifier,
+        child: InheritedProvider<int>(
+          update: update,
+          builder: builder,
+          child: child,
+        ),
+      ),
+    );
+    expect(find.text('notifier: 1 value: 1'), findsOneWidget);
+    verify(builder(argThat(isNotNull), 1, child)).called(1);
+    verifyNoMoreInteractions(builder);
+    verify(update(any, any)).called(1);
+    verifyNoMoreInteractions(update);
+  });
+  testWidgets('provider.of throws if listen:true outside of the widget tree',
+      (tester) async {
     await tester.pumpWidget(
       InheritedProvider<int>.value(
         value: 42,
@@ -242,7 +303,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
   });
 
   group('InheritedProvider.value()', () {
-    testWidgets('markNeedsNotifyDependents during startListening is noop', (tester) async {
+    testWidgets('markNeedsNotifyDependents during startListening is noop',
+        (tester) async {
       await tester.pumpWidget(
         InheritedProvider<int>.value(
           value: 42,
@@ -254,7 +316,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
         ),
       );
     });
-    testWidgets('startListening called again when create returns new value', (tester) async {
+    testWidgets('startListening called again when create returns new value',
+        (tester) async {
       final stopListening = StopListeningMock();
       final startListening = StartListeningMock<int>(stopListening);
 
@@ -481,7 +544,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
     });
   });
   group('InheritedProvider()', () {
-    testWidgets("provider notifying dependents doesn't call update", (tester) async {
+    testWidgets("provider notifying dependents doesn't call update",
+        (tester) async {
       final notifier = ValueNotifier(0);
       final mock = ValueBuilderMock<ValueNotifier<int>>(notifier);
 
@@ -526,7 +590,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       expect(find.text('42'), findsOneWidget);
     });
 
-    testWidgets('update lazy loaded can call Provider.of with listen:true', (tester) async {
+    testWidgets('update lazy loaded can call Provider.of with listen:true',
+        (tester) async {
       await tester.pumpWidget(
         InheritedProvider<int>.value(
           value: 42,
@@ -539,7 +604,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
 
       expect(Provider.of<String>(context, listen: false), equals('42'));
     });
-    testWidgets('markNeedsNotifyDependents during startListening is noop', (tester) async {
+    testWidgets('markNeedsNotifyDependents during startListening is noop',
+        (tester) async {
       await tester.pumpWidget(
         InheritedProvider<int>(
           update: (_, __) => 24,
@@ -655,7 +721,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       verifyNoMoreInteractions(stopListening);
     });
 
-    testWidgets('startListening called again when create returns new value', (tester) async {
+    testWidgets('startListening called again when create returns new value',
+        (tester) async {
       final stopListening = StopListeningMock();
       final startListening = StartListeningMock<int>(stopListening);
 
@@ -779,7 +846,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
           },
         );
         final update = ValueBuilderMock<int>();
-        when(update(any, any)).thenAnswer((i) => (i.positionalArguments[1] as int) * 2);
+        when(update(any, any))
+            .thenAnswer((i) => (i.positionalArguments[1] as int) * 2);
 
         await tester.pumpWidget(
           InheritedProvider<int>(
@@ -1022,7 +1090,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       verify(dispose(context, 42)).called(1);
       verifyNoMoreInteractions(dispose);
     });
-    testWidgets('builder unmount, dispose not called if value never read', (tester) async {
+    testWidgets('builder unmount, dispose not called if value never read',
+        (tester) async {
       final dispose = DisposeMock<int>();
 
       await tester.pumpWidget(
@@ -1146,7 +1215,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
     // TODO: stopListening cannot be null
     testWidgets('startListening', (tester) async {
       final stopListening = StopListeningMock();
-      final startListening = DeferredStartListeningMock<ValueNotifier<int>, int>(
+      final startListening =
+          DeferredStartListeningMock<ValueNotifier<int>, int>(
         (e, setState, controller, value) {
           setState(controller.value);
           return stopListening;
@@ -1222,8 +1292,10 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
 
       expect(tester.takeException(), isAssertionError);
     });
-    testWidgets("startListening doesn't need setState if already initialized", (tester) async {
-      final startListening = DeferredStartListeningMock<ValueNotifier<int>, int>(
+    testWidgets("startListening doesn't need setState if already initialized",
+        (tester) async {
+      final startListening =
+          DeferredStartListeningMock<ValueNotifier<int>, int>(
         (e, setState, controller, value) {
           setState(controller.value);
           return () {};
@@ -1241,7 +1313,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
 
       expect(find.text('0'), findsOneWidget);
 
-      final startListening2 = DeferredStartListeningMock<ValueNotifier<int>, int>();
+      final startListening2 =
+          DeferredStartListeningMock<ValueNotifier<int>, int>();
       when(startListening2(any, any, any, any)).thenReturn(() {});
       final controller2 = ValueNotifier<int>(0);
 
@@ -1355,8 +1428,10 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       expect(find.text('1'), findsOneWidget);
       expect(buildCount, equals(2));
     });
-    testWidgets('startListening never leave the widget uninitialized', (tester) async {
-      final startListening = DeferredStartListeningMock<ValueNotifier<int>, int>();
+    testWidgets('startListening never leave the widget uninitialized',
+        (tester) async {
+      final startListening =
+          DeferredStartListeningMock<ValueNotifier<int>, int>();
       when(startListening(any, any, any, any)).thenReturn(() {});
       final controller = ValueNotifier<int>(0);
 
@@ -1375,7 +1450,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       );
     });
 
-    testWidgets('startListening called again on controller change', (tester) async {
+    testWidgets('startListening called again on controller change',
+        (tester) async {
       var buildCount = 0;
       final child = Consumer<int>(builder: (_, value, __) {
         buildCount++;
@@ -1383,7 +1459,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       });
 
       final stopListening = StopListeningMock();
-      final startListening = DeferredStartListeningMock<ValueNotifier<int>, int>(
+      final startListening =
+          DeferredStartListeningMock<ValueNotifier<int>, int>(
         (e, setState, controller, value) {
           setState(controller.value);
           return stopListening;
@@ -1405,7 +1482,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       verifyZeroInteractions(stopListening);
 
       final stopListening2 = StopListeningMock();
-      final startListening2 = DeferredStartListeningMock<ValueNotifier<int>, int>(
+      final startListening2 =
+          DeferredStartListeningMock<ValueNotifier<int>, int>(
         (e, setState, controller, value) {
           setState(controller.value);
           return stopListening2;
@@ -1485,7 +1563,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       expect(of<int>(), equals(0));
 
       verify(create(argThat(isNotNull))).called(1);
-      verify(startListening(argThat(isNotNull), argThat(isNotNull), '0', null)).called(1);
+      verify(startListening(argThat(isNotNull), argThat(isNotNull), '0', null))
+          .called(1);
 
       expect(of<int>(), equals(0));
 
@@ -1667,7 +1746,8 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
     verify(startListening(argThat(isNotNull), 42)).called(1);
   });
 
-  testWidgets('InheritedProvider.value lazy loading can be disabled', (tester) async {
+  testWidgets('InheritedProvider.value lazy loading can be disabled',
+      (tester) async {
     final startListening = StartListeningMock<int>(() {});
 
     await tester.pumpWidget(
@@ -1702,8 +1782,10 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       verifyNoMoreInteractions(create);
     },
   );
-  testWidgets('DeferredInheritedProvider lazy loading can be disabled', (tester) async {
-    final startListening = DeferredStartListeningMock<int, int>((a, setState, c, d) {
+  testWidgets('DeferredInheritedProvider lazy loading can be disabled',
+      (tester) async {
+    final startListening =
+        DeferredStartListeningMock<int, int>((a, setState, c, d) {
       setState(0);
       return () {};
     });
@@ -1718,11 +1800,14 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       ),
     );
 
-    verify(startListening(argThat(isNotNull), argThat(isNotNull), 42, null)).called(1);
+    verify(startListening(argThat(isNotNull), argThat(isNotNull), 42, null))
+        .called(1);
     verifyNoMoreInteractions(startListening);
   });
-  testWidgets('DeferredInheritedProvider.value lazy loading can be disabled', (tester) async {
-    final startListening = DeferredStartListeningMock<int, int>((a, setState, c, d) {
+  testWidgets('DeferredInheritedProvider.value lazy loading can be disabled',
+      (tester) async {
+    final startListening =
+        DeferredStartListeningMock<int, int>((a, setState, c, d) {
       setState(0);
       return () {};
     });
@@ -1737,13 +1822,205 @@ DeferredInheritedProvider<int, int>(controller: 42, value: 24)'''),
       ),
     );
 
-    verify(startListening(argThat(isNotNull), argThat(isNotNull), 42, null)).called(1);
+    verify(startListening(argThat(isNotNull), argThat(isNotNull), 42, null))
+        .called(1);
     verifyNoMoreInteractions(startListening);
+  });
+
+  testWidgets('selector', (tester) async {
+    final notifier = ValueNotifier(0);
+    var buildCount = 0;
+
+    await tester.pumpWidget(
+      ChangeNotifierProvider(
+        create: (_) => notifier,
+        child: Builder(builder: (context) {
+          buildCount++;
+          final isEven =
+              context.select((ValueNotifier<int> value) => value.value.isEven);
+
+          return Text('$isEven', textDirection: TextDirection.ltr);
+        }),
+      ),
+    );
+
+    expect(buildCount, 1);
+    expect(find.text('true'), findsOneWidget);
+
+    notifier.value = 1;
+    await tester.pump();
+
+    expect(buildCount, 2);
+    expect(find.text('false'), findsOneWidget);
+
+    notifier.value = 3;
+    await tester.pump();
+
+    expect(buildCount, 2);
+    expect(find.text('false'), findsOneWidget);
+  });
+
+  testWidgets('can select multiple types from same provider', (tester) async {
+    var buildCount = 0;
+
+    var builder = Builder(builder: (context) {
+      buildCount++;
+      final isNotNull = context.select((int value) => value != null);
+      final isAbove0 = context
+          .select((int value) => (value == null || value > 0).toString());
+
+      return Text('$isNotNull $isAbove0', textDirection: TextDirection.ltr);
+    });
+
+    await tester.pumpWidget(Provider.value(value: 0, child: builder));
+
+    expect(buildCount, 1);
+    expect(find.text('true false'), findsOneWidget);
+
+    await tester.pumpWidget(Provider.value(value: -1, child: builder));
+
+    expect(buildCount, 1);
+
+    await tester.pumpWidget(Provider.value(value: 1, child: builder));
+
+    expect(buildCount, 2);
+    expect(find.text('true true'), findsOneWidget);
+
+    await tester.pumpWidget(Provider<int>.value(value: null, child: builder));
+
+    expect(buildCount, 3);
+    expect(find.text('false true'), findsOneWidget);
+  });
+  testWidgets('can select same type on two different providers',
+      (tester) async {
+    var buildCount = 0;
+
+    var builder = Builder(builder: (context) {
+      buildCount++;
+      final intValue = context.select((int value) => value.toString());
+      final stringValue = context.select((String value) => value);
+
+      return Text('$intValue $stringValue', textDirection: TextDirection.ltr);
+    });
+
+    await tester.pumpWidget(
+      MultiProvider(
+        providers: [
+          Provider.value(value: 0),
+          Provider.value(value: 'a'),
+        ],
+        child: builder,
+      ),
+    );
+
+    expect(buildCount, 1);
+    expect(find.text('0 a'), findsOneWidget);
+
+    await tester.pumpWidget(
+      MultiProvider(
+        providers: [
+          Provider.value(value: 0),
+          Provider.value(value: 'a'),
+        ],
+        child: builder,
+      ),
+    );
+
+    expect(buildCount, 1);
+
+    await tester.pumpWidget(
+      MultiProvider(
+        providers: [
+          Provider.value(value: 1),
+          Provider.value(value: 'a'),
+        ],
+        child: builder,
+      ),
+    );
+
+    expect(buildCount, 2);
+    expect(find.text('1 a'), findsOneWidget);
+  });
+  testWidgets(
+      'can give a key to selectors to select same type twice on same provider',
+      (tester) async {
+    var buildCount = 0;
+    var child = Builder(builder: (context) {
+      buildCount++;
+      final value = context.select((int value) => value.isEven, 0);
+      final value2 = context.select((int value) => value.isNegative, 1);
+
+      return Text('$value $value2', textDirection: TextDirection.ltr);
+    });
+
+    await tester.pumpWidget(Provider.value(value: 0, child: child));
+
+    expect(find.text('true false'), findsOneWidget);
+    expect(buildCount, 1);
+
+    await tester.pumpWidget(Provider.value(value: 2, child: child));
+
+    expect(find.text('true false'), findsOneWidget);
+    expect(buildCount, 1);
+
+    await tester.pumpWidget(Provider.value(value: -2, child: child));
+
+    expect(find.text('true true'), findsOneWidget);
+    expect(buildCount, 2);
+
+    await tester.pumpWidget(Provider.value(value: -4, child: child));
+
+    expect(find.text('true true'), findsOneWidget);
+    expect(buildCount, 2);
+
+    await tester.pumpWidget(Provider.value(value: -3, child: child));
+
+    expect(find.text('false true'), findsOneWidget);
+    expect(buildCount, 3);
+
+    await tester.pumpWidget(Provider.value(value: -2, child: child));
+
+    expect(find.text('true true'), findsOneWidget);
+    expect(buildCount, 4);
+  });
+  testWidgets(
+      'selecting same provider + same type twice is not allowed without keys',
+      (tester) async {
+    await tester.pumpWidget(
+      MultiProvider(
+        providers: [
+          Provider(create: (_) => 42),
+        ],
+        child: Builder(builder: (context) {
+          final value = context.select((int value) => value);
+          final value2 = context.select((int value) => value);
+
+          return Text('$value $value2', textDirection: TextDirection.ltr);
+        }),
+      ),
+    );
+
+    expect(tester.takeException(), isAssertionError);
   });
 }
 
+class Model {
+  int a;
+  String b;
+}
+
+class Example extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    final a = context.select((Model model) => model.a);
+    final b = context.select((Model model) => model.b);
+    return Text('$a $b');
+  }
+}
+
 class Test<T> extends StatefulWidget {
-  const Test({Key key, this.didChangeDependencies, this.build}) : super(key: key);
+  const Test({Key key, this.didChangeDependencies, this.build})
+      : super(key: key);
 
   final ValueBuilderMock<T> didChangeDependencies;
   final ValueBuilderMock<T> build;
@@ -1755,7 +2032,8 @@ class Test<T> extends StatefulWidget {
 class _TestState<T> extends State<Test<T>> {
   @override
   void didChangeDependencies() {
-    widget.didChangeDependencies?.call(this.context, Provider.of<T>(this.context));
+    widget.didChangeDependencies
+        ?.call(this.context, Provider.of<T>(this.context));
     super.didChangeDependencies();
   }